+++ cf_xarray/tests/test_accessor.py
@@ -11,8 +11,17 @@ from xarray.testing import assert_allclose, assert_identical
 
 import cf_xarray  # noqa
 
+from ..datasets import (
+    airds,
+    anc,
+    ds_no_attrs,
+    forecast,
+    mollwds,
+    multiple,
+    popds,
+    romsds,
+)
 from . import raise_if_dask_computes
-from .datasets import airds, anc, ds_no_attrs, multiple, popds
 
 mpl.use("Agg")
 
@@ -235,7 +244,6 @@ def test_rename_like():
                 reason="xarray GH4120. any test after this will fail since attrs are lost"
             ),
         ),
-        # groupby("time.day")?
     ),
 )
 def test_wrapped_classes(obj, attr, xrkwargs, cfkwargs):
@@ -365,15 +373,20 @@ def test_dataarray_getitem():
     assert_identical(air.cf["area_grid_cell"], air.cell_area.reset_coords(drop=True))
 
 
-@pytest.mark.parametrize("obj", dataarrays)
-def test_dataarray_plot(obj):
+def test_dataarray_plot():
 
-    rv = obj.isel(time=1).cf.plot(x="X", y="Y")
+    obj = airds.air
+
+    rv = obj.isel(time=1).transpose("lon", "lat").cf.plot()
     assert isinstance(rv, mpl.collections.QuadMesh)
+    assert all(v > 180 for v in rv.axes.get_xlim())
+    assert all(v < 200 for v in rv.axes.get_ylim())
     plt.close()
 
-    rv = obj.isel(time=1).cf.plot.contourf(x="X", y="Y")
+    rv = obj.isel(time=1).transpose("lon", "lat").cf.plot.contourf()
     assert isinstance(rv, mpl.contour.QuadContourSet)
+    assert all(v > 180 for v in rv.axes.get_xlim())
+    assert all(v < 200 for v in rv.axes.get_ylim())
     plt.close()
 
     rv = obj.cf.plot(x="X", y="Y", col="T")
@@ -386,6 +399,29 @@ def test_dataarray_plot(obj):
 
     rv = obj.isel(lat=[0, 1], lon=1).cf.plot.line(x="T", hue="Y")
     assert all([isinstance(line, mpl.lines.Line2D) for line in rv])
+    plt.close()
+
+    # set y automatically
+    rv = obj.isel(time=0, lon=1).cf.plot.line()
+    np.testing.assert_equal(rv[0].get_ydata(), obj.lat.data)
+    plt.close()
+
+    # don't set y automatically
+    rv = obj.isel(time=0, lon=1).cf.plot.line(x="lat")
+    np.testing.assert_equal(rv[0].get_xdata(), obj.lat.data)
+    plt.close()
+
+    # various line plots and automatic guessing
+    rv = obj.cf.isel(T=1, Y=[0, 1, 2]).cf.plot.line()
+    np.testing.assert_equal(rv[0].get_xdata(), obj.lon.data)
+    plt.close()
+
+    # rv = obj.cf.isel(T=1, Y=[0, 1, 2]).cf.plot(hue="Y")
+    # np.testing.assert_equal(rv[0].get_xdata(), obj.lon.data)
+    # plt.close()
+
+    rv = obj.cf.isel(T=1, Y=[0, 1, 2]).cf.plot.line()
+    np.testing.assert_equal(rv[0].get_xdata(), obj.lon.data)
     plt.close()
 
     obj = obj.copy(deep=True)
@@ -567,6 +603,15 @@ def test_bounds_to_vertices():
     assert "time_bounds" in dsc
 
 
+def test_get_bounds_dim_name():
+    ds = airds.copy(deep=True).cf.add_bounds("lat")
+    assert ds.cf.get_bounds_dim_name("latitude") == "bounds"
+    assert ds.cf.get_bounds_dim_name("lat") == "bounds"
+
+    assert mollwds.cf.get_bounds_dim_name("longitude") == "bounds"
+    assert mollwds.cf.get_bounds_dim_name("lon") == "bounds"
+
+
 def test_docstring():
     assert "One of ('X'" in airds.cf.groupby.__doc__
     assert "One or more of ('X'" in airds.cf.mean.__doc__
@@ -658,7 +703,7 @@ def test_missing_variable_in_coordinates():
 
 
 def test_Z_vs_vertical_ROMS():
-    from .datasets import romsds
+    from ..datasets import romsds
 
     assert_identical(romsds.s_rho.reset_coords(drop=True), romsds.temp.cf["Z"])
     assert_identical(
@@ -684,8 +729,6 @@ def test_Z_vs_vertical_ROMS():
 
 
 def test_param_vcoord_ocean_s_coord():
-    from .datasets import romsds
-
     romsds.s_rho.attrs["standard_name"] = "ocean_s_coordinate_g2"
     Zo_rho = (romsds.hc * romsds.s_rho + romsds.Cs_r * romsds.h) / (
         romsds.hc + romsds.h
@@ -788,3 +831,63 @@ def test_drop_dims(ds):
     # Axis and coordinate
     for cf_name in ["X", "longitude"]:
         assert_identical(ds.drop_dims("lon"), ds.cf.drop_dims(cf_name))
+
+
+def test_new_standard_name_mappers():
+    assert_identical(forecast.cf.mean("realization"), forecast.mean("M"))
+    assert_identical(
+        forecast.cf.mean(["realization", "forecast_period"]), forecast.mean(["M", "L"])
+    )
+    assert_identical(forecast.cf.chunk({"realization": 1}), forecast.chunk({"M": 1}))
+    assert_identical(forecast.cf.isel({"realization": 1}), forecast.isel({"M": 1}))
+    assert_identical(forecast.cf.isel(**{"realization": 1}), forecast.isel(**{"M": 1}))
+    assert_identical(
+        forecast.cf.groupby("forecast_reference_time.month").mean(),
+        forecast.groupby("S.month").mean(),
+    )
+
+
+def test_possible_x_y_plot():
+    from ..accessor import _possible_x_y_plot
+
+    # choose axes
+    assert _possible_x_y_plot(airds.air.isel(time=1), "x") == "lon"
+    assert _possible_x_y_plot(airds.air.isel(time=1), "y") == "lat"
+    assert _possible_x_y_plot(airds.air.isel(lon=1), "y") == "lat"
+    assert _possible_x_y_plot(airds.air.isel(lon=1), "x") == "time"
+
+    # choose coordinates over axes
+    assert _possible_x_y_plot(popds.UVEL, "x") == "ULONG"
+    assert _possible_x_y_plot(popds.UVEL, "y") == "ULAT"
+    assert _possible_x_y_plot(popds.TEMP, "x") == "TLONG"
+    assert _possible_x_y_plot(popds.TEMP, "y") == "TLAT"
+
+    assert _possible_x_y_plot(popds.UVEL.drop_vars("ULONG"), "x") == "nlon"
+
+    # choose X over T, Y over Z
+    def makeds(*dims):
+        coords = {dim: (dim, np.arange(3), {"axis": dim}) for dim in dims}
+        return xr.DataArray(np.zeros((3, 3)), dims=dims, coords=coords)
+
+    yzds = makeds("Y", "Z")
+    assert _possible_x_y_plot(yzds, "y") == "Z"
+    assert _possible_x_y_plot(yzds, "x") is None
+
+    xtds = makeds("X", "T")
+    assert _possible_x_y_plot(xtds, "y") is None
+    assert _possible_x_y_plot(xtds, "x") == "X"
+
+
+def test_groupby_special_ops():
+    cfgrouped = airds.cf.groupby_bins("latitude", np.arange(20, 50, 10))
+    grouped = airds.groupby_bins("lat", np.arange(20, 50, 10))
+
+    # __iter__
+    for (label, group), (cflabel, cfgroup) in zip(grouped, cfgrouped):
+        assert label == cflabel
+        assert_identical(group, cfgroup)
+
+    # arithmetic
+    expected = grouped - grouped.mean()
+    actual = grouped - cfgrouped.mean()
+    assert_identical(expected, actual)